fix(ai): allow moved safety guard lines
